Bugs fixed

+++ types.h
@@ -4,8 +4,8 @@
 typedef __m128i *contr;
 typedef uint32_t meter;
 typedef uint16_t place;
-typedef uint16_t agent;
+typedef uint32_t agent;
 typedef uint32_t penny;
-typedef uint16_t edge;
+typedef uint32_t edge;
 typedef uint32_t id;
 typedef float dist;
